requirements.txt
aws-cdk-lib==2.151.0
constructs>=10.0.0,<11.0.0
pydantic==2.8.2
pydantic_settings==2.4.0
